Animation and Movement Overhaul. Triggerzones
The new Movement script can be applied now be applied to both players and enemies, completely without errors. Meaning, making a new enemy takes about 30 seconds. The scripts changes the tag of the object corresponding to its physical state such as Running, Jumping and being idle. The new animator script that can also be applied to both the player and enemies uses the tag to decide which animation to play. For example if the player tag is "Running" then the animator will play the run animation. To make the animation, you just drag the different sprites into corresponding frame in the corresponding sprite index. The new movement script checks the layer of the object that it is placed on to decide how to move the object. I've made an empty game object that we can child triggerzones to for organization.